requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
dateparser==1.1.8
Django==3.2.11
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
drf-yasg==1.21.5
frozenlist==1.3.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
multidict==5.2.0
packaging==23.0
Pillow==9.4.0
PyJWT==2.6.0
pyrsistent==0.19.3
python-binance==1.0.16
python-dateutil==2.8.2
python-decouple==3.7
python-dotenv==1.0.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2021.10.23
requests==2.26.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
typing-extensions==3.10.0.2
tzdata==2022.1
tzlocal==4.3
ujson==4.2.0
uritemplate==4.1.1
urllib3==1.25.11
websockets==8.1
yarl==1.6.3
